+++ libs/core/src/handlers/commands.ts
@@ -75,6 +75,7 @@ export enum SeelenCommand {
   GetUserFolderContent = "get_user_folder_content",
   GetUserAppWindows = "get_user_app_windows",
   GetUserAppWindowsPreviews = "get_user_app_windows_previews",
+  GetUserAppWindowsColors = "get_user_app_windows_colors",
   GetMediaDevices = "get_media_devices",
   GetMediaSessions = "get_media_sessions",
   MediaPrev = "media_prev",