+++ libs/core/src/handlers/events.rs
@@ -59,6 +59,7 @@ slu_events_declaration! {
     UserFolderChanged(FolderChangedArgs) as "user::known-folder-changed",
     UserAppWindowsChanged(Vec<UserAppWindow>) as "user::windows-changed",
     UserAppWindowsPreviewsChanged(HashMap<isize, UserAppWindowPreview>) as "user::windows-previews-changed",
+    UserAppWindowsColorsChanged(HashMap<isize, UserAppWindowColors>) as "user::windows-colors-changed",
 
     MediaSessions(Vec<MediaPlayer>) as "media-sessions",
     MediaDevices([Vec<MediaDevice>; 2]) as "media::devices",